Updated Review page.
Click on the picture to continue forward through the slides upon failure, then answer the quiz questions correctly to return to the video.

To add new slides you will have to add the pictures to a folder, copy one of the sample divs for the pictures and change the id to something unique, then add that unique id to the idArray in the javascript portion.

Good Luck,
~Scott M.